Avoid deleting existing Frogbot comments
