+++ docs/install-gitlab.md
@@ -110,6 +110,10 @@ frogbot-scan:
     # Displays all existing vulnerabilities, including the ones that were added by the pull request.
     # JF_INCLUDE_ALL_VULNERABILITIES: "TRUE"
 
+    # [Optional, default: "FALSE"]
+    # Keep old comments that were added by previous Frogbot scans on the pull request.
+    # JF_AVOID_OLD_PR_COMMENTS_DELETION: "TRUE"
+
     # [Optional, default: "TRUE"]
     # Fails the Frogbot task if any security issue is found.
     # JF_FAIL: "FALSE"
