+++ docs/templates/jfrog-pipelines/pipelines-npm.yml
@@ -126,6 +126,10 @@ pipelines:
             # Displays all existing vulnerabilities, including the ones that were added by the pull request.
             # JF_INCLUDE_ALL_VULNERABILITIES: "TRUE"
 
+            # [Optional, default: "FALSE"]
+            # Keep old comments that were added by previous Frogbot scans on the pull request.
+            # JF_AVOID_OLD_PR_COMMENTS_DELETION: "TRUE"
+
             # [Optional, default: "TRUE"]
             # Fails the Frogbot task if any security issue is found.
             # JF_FAIL: "FALSE"
@@ -139,18 +143,6 @@ pipelines:
             # [Optional, default: "."]
             # Relative path to the project in the git repository
             # JF_WORKING_DIR: path/to/project/dir
-
-            # [Optional]
-            # Xray Watches. Learn more about them here: https://www.jfrog.com/confluence/display/JFROG/Configuring+Xray+Watches
-            # JF_WATCHES: <watch-1>,<watch-2>...<watch-n>
-
-            # [Optional, default: "FALSE"]
-            # Displays all existing vulnerabilities, including the ones that were added by the pull request.
-            # JF_INCLUDE_ALL_VULNERABILITIES: "TRUE"
-
-            # [Optional, default: "TRUE"]
-            # Fails the Frogbot task if any security issue is found.
-            # JF_FAIL: "FALSE"
 
             # [Optional]
             # Template for the branch name generated by Frogbot when creating pull requests with fixes.
